srv_time/requirements.txt
grpcio==1.43.0
grpcio-tools==1.43.0
pytz==2022.6